document coldstart drivers incompatible with paralllel cth
cth now runs tests in parallel by default, and this completely breaks coldstart and doh-coldstart, which are simple reference/dev/debugging drivers. It is still possible to write tests for them and run them with cth, but you have to run cth -j 1. If you don't do this, the tests will simply fail.