Remove cruft
- removed tools/pldoh submodule (not used in test automation; obsoleted by Javascript test library development)
- removed old_tests/* (obsoleted by tests/)
- removed drivers/doh-perma

pldoh will continue to be developed as I have time, but it's going to be continually
broken by improvements to the DoH ABIs. It's a tool that's going to be helpful for
on-line testing, mostly (and playing the game, and writing good bots).

The Javascript test library (tools/dohtestlib), on the other hand, will set the DoH test
tree on fire when the DoH ABIs break it. That will be a case of "you break it, you fix it".
Can't merge to main when your PR's CI results are red.